edits to dataset

+++ MLP/mlp_architecture.py
@@ -23,8 +23,6 @@ class MLPModel(nn.Module):
         self.output = nn.Softmax(dim=1)
         
     def forward(self, x):
-        #RuntimeError: mat1 and mat2 shapes cannot be multiplied (3200x220500 and 100x20)
-        #44100*5 = 220500
         x = self.input(x)
         
         for hidden_layer in self.hidden_layers:
@@ -35,13 +33,12 @@ class MLPModel(nn.Module):
 
 
 class MLPDataset(torch.utils.data.Dataset):
-    #TODO: Randomize the gun sample in AudioSampler
     data = AudioSampler('city.wav','kaggle_sounds/Zastava M92/9 (1).wav')
 
     def __init__(self,hyperparams,size):
         self.window_size = hyperparams['window_size']
         self.X,self.y = MLPDataset.data.sample_array(size,self.window_size, convert_to_mono=True)
-        self.X = torch.tensor(self.X).float()
+        self.X = torch.tensor(self.X)
         self.y = torch.tensor(self.y)
     
     def __len__(self):
